requirements.txt
fastapi==0.109.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
pymysql==1.1.0
python-dotenv
python-jose
passlib
python-multipart